+++ simulation/ninja/Ninja.cpp
@@ -230,7 +230,7 @@ void Ninja::PostCollision(Simulator* sim) {
             wallList_Y.clear();
             result_logical.Clear();
             sim->objGrid->GatherCellContentsInNeighbourhood(pos, objList);
-            for (int i = 0; i < objList.size(); ++i) {
+            for (size_t i = 0; i < objList.size(); ++i) {
                 Entity_Base* entity = objList[i];
                 if (entity->CollideVsCircle_Logical(sim, this, result_logical, pos, vel, oldpos, r, epsilon)) {
                     if (result_logical.vec_y == 0) {
@@ -241,7 +241,7 @@ void Ninja::PostCollision(Simulator* sim) {
             }
             double effectiveRadius = r + epsilon;
             sim->segGrid->GatherCellContentsFromWorldspaceRegion(pos.x - effectiveRadius, pos.y - effectiveRadius, pos.x + effectiveRadius, pos.y + effectiveRadius, segList);
-            for (int i = 0; i < segList.size(); ++i) {
+            for (size_t i = 0; i < segList.size(); ++i) {
                 segList[i]->GetClosestPoint(pos, seg_cp);
                 double dx = pos.x - seg_cp.x;
                 double dy = pos.y - seg_cp.y;
@@ -331,7 +331,7 @@ void Ninja::CollideVsObjects(Simulator* sim) {
         result_physical.Clear();
         sim->objGrid->GatherCellContentsInNeighbourhood(pos, objList);
 
-        for (int i = 0; i < objList.size(); ++i) {
+        for (size_t i = 0; i < objList.size(); ++i) {
             Entity_Base* entity = objList[i];
             
             if (entity->CollideVsCircle_Physical(result_physical, pos, vel, oldpos, r)) {
